Split things up and tidy.
